Create patch set 3
Uploaded patch set 3.

Patch-set: 3
Subject: Update logical network elements references to each other on a change. When an element is changed its references should be updated as well. For example, if a logical router is deleted, then the ports that belong to that router should be removed as well.
Commit: 459dde705d49797128e3bb765e2aa55366096d5a