Update patch set 3
Patch Set 3:

> 1) why remove parent? 

The parent here is useless and isn't used. So removing it is cleaning up the repo

> 2) why remove style.xml? 

I'm not exactly sure what that is, since it seems related to eclipse configuration. If it is the case, I think it shouldn't be in the upstream repo as this is really local, enduser configuration (but that's only my vision, so doesn't import that much :) )
Else if this is checkstyle related, it should be removed to use odlparent's one.

Patch-set: 3