Convert to blueprint
Converted all the bundles that were using the config subsystem to
blueprint as the config subsystem is deprecated in Nitrogen with
likely removal in Flourine.

The fabric-mgr-impl and uln-mapper-impl bundles were refactored
a bit to eliminate the use of statics for wiring which makes it
cleaner. Some new interfaces and OSGi services were added.

Change-Id: I524513a07ccbdcbcceabd0798ed92491d80f40f3
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>